Media Present on BUILTIN_SDCARD
Found the Media Present when card pulled out on T4.1 was taking about 3 seconds to complete.

So this version tries to use the DAT3 pin on the SDCard to detect when the card is inserted.  Only if this goes high do we then try to initialize the SDCard again.

there is an elapsedMicros that I use after this to try for up to a few seconds after this, before I set DAT3 back to IO pin... This allows the SDCard time to initialize...